chore: reorganize project structure and update .gitignore
