--- llama3-vertexai-finetuning/.gitignore
@@ -1,3 +1,0 @@
-credentials/*
-__pycache__/*
-*.pycllama3-vertexai-finetuning/setup_credentials.py
