--- llama3-vertexai-finetuning/requirements.txt
@@ -1,33 +1,0 @@
-# Core dependencies
-google-cloud-aiplatform>=1.30.0
-vertexai>=0.3.0
-datasets>=2.13.0
-pandas>=2.0.0
-numpy>=1.24.0
-
-# Visualization dependencies
-streamlit>=1.24.0
-plotly>=5.14.0
-matplotlib>=3.7.0
-seaborn>=0.12.0
-
-# Web server
-flask>=2.3.0
-flask-cors>=4.0.0
-gunicorn>=21.0.0
-requests>=2.31.0
-
-# Google Cloud dependencies
-google-auth>=2.22.0
-google-cloud-storage>=2.10.0
-
-# Utilities
-tqdm>=4.65.0
-python-dotenv>=1.0.0
-pyyaml>=6.0
-
-
-bert_score>=0.3.12
-datasets>=2.13.0
-sentence-transformers>=2.2.2
-matplotlib>=3.7.0
